* New upstream release:
    - introduce new branch attribute "nolwh."
    - refine parsing options at mount/remount time, reported by Sandino
      Flores Moreno.
      + remove whiteout-base and whplink-dir in nopllink and br_mod(),
        but br_del().
      + increment sbinfo generation in a new function
        aufs_remount_fs().
      + support multiple xino/noxino options in another way.
      + force "ro" when unknown branch permission is specified.
      + at mount-time, interpret options in this order.
              * simple flags, except xino/noxino, udba=inotify and dlgt.
              * branches.
              * xino/noxino.
              * udba=inotify.
              * dlgt.
      + at remount-time,
              * interpret options in the given order left to right,
                except dlgt which is disabled in interpreting.
              * create whiteout-base and whplink-dir if neccessary.
              * dlgt if neccessary.
      + new functions, clean_wh(), clean_plink(), test_linkable(),
        plink_dir().
    - support 64bit environtment, reported by Vasily Tarasov.
    - introduce a sample for kprobe.
    - bugfix: do not traverse the sub-mounts under aufs at refreshing
      cached dirs after remount.
    - bugfix: simultaneous read in 'dlgt' mode, reported and fixed by
      Ulrich Holeschak.
    - bugfix: LINUX_VERSION_CODE for linux-2.6.16.
    - still testing inotify.
    - support lockdep in nfsd, still testing.
    - minor optimization for dput().
    - new test functions br_writable(), br_whable() and br_linkable_wh(),
      replacing au_is_whable().
    - do write_lock() in reinit_br_wh().
    - refine updating /sys/fs/aufs/stat.
    - wkq_nowait() calls schedule_work().
    - bugfix: forgot dput in a error case.
    - introduce wkq_nowait().
    - refine udba=inotify, still testing.
    - extract some functions from sbinfo.c, create a new file plink.c.
  * add myself to uploaders.

 -- Stefan Lippers-Hollmann <s.l-h@gmx.de>  Sat, 14 Apr 2007 01:50:13 +0200

+++ fs/aufs/f_op.c
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-/* $Id: f_op.c,v 1.21 2007/03/27 12:49:00 sfjro Exp $ */
+/* $Id: f_op.c,v 1.22 2007/04/09 02:44:47 sfjro Exp $ */
 
 #include <linux/fsnotify.h>
 #include <linux/pagemap.h>
@@ -130,7 +130,8 @@ static ssize_t aufs_read(struct file *file, char __user *buf, size_t count,
 	int dlgt;
 
 	dentry = file->f_dentry;
-	LKTRTrace("%.*s, cnt %d, pos %Ld\n", DLNPair(dentry), count, *ppos);
+	LKTRTrace("%.*s, cnt %lu, pos %Ld\n",
+		  DLNPair(dentry), (unsigned long)count, *ppos);
 
 	sb = dentry->d_sb;
 	si_read_lock(sb);
@@ -177,7 +178,8 @@ static ssize_t aufs_write(struct file *file, const char __user *__buf,
 	int dlgt;
 
 	dentry = file->f_dentry;
-	LKTRTrace("%.*s, cnt %d, pos %Ld\n", DLNPair(dentry), count, *ppos);
+	LKTRTrace("%.*s, cnt %lu, pos %Ld\n",
+		  DLNPair(dentry), (unsigned long)count, *ppos);
 
 	inode = dentry->d_inode;
 	i_lock(inode);
@@ -467,8 +469,8 @@ static ssize_t aufs_sendfile(struct file *file, loff_t *ppos,
 	struct super_block *sb;
 
 	dentry = file->f_dentry;
-	LKTRTrace("%.*s, pos %Ld, cnt %d, loopback %d\n",
-		  DLNPair(dentry), *ppos, count, loopback);
+	LKTRTrace("%.*s, pos %Ld, cnt %lu, loopback %d\n",
+		  DLNPair(dentry), *ppos, (unsigned long)count, loopback);
 
 	sb = dentry->d_sb;
 	si_read_lock(sb);
